requirements.txt
pytorch
numpy
scipy
matplotlib